* commit '11e252a82ce9458b2630a692eb80128557291803':
  [GTSAVTAMER-8677] Update positioning
  [GTSAVTAMER-8677] Remove tail from quote update id, all 16 characters will be used
  efcSetSessionCntr()
  sw fixes
  QED SW support
  ../new_api_tests/efcFastSweepFireTestNew.c
  strat ver
  reports
  compiling
  test
  b4 test
  qed ehp
  tab
  EkaEpmRegion fixed action budgetting
  new APIs + tests
  efcInit() to initialize EpmTx if needed. compat/EfcMultiStrategies.h
  [GTSAVTAMER-8677] Revert "[GTSAVTAMER-8677] Add logging"
  [GTSAVTAMER-8677] Add logging

+++ apisrc/Ehp/EhpNom.c
@@ -33,7 +33,7 @@ int EhpNom::init() {
 
 
 int EhpNom::createAddOrderShort() {
-  uint8_t msgId   = 'a';
+  uint16_t msgId   = 'a';
   int     msgType = AddOrderShortMsg;
 
   conf.params.bytes4Strategy[msgType].msgId    = msgId;
@@ -124,7 +124,7 @@ int EhpNom::createAddOrderShort() {
 }
 
 int EhpNom::createAddOrderLong() {
-  uint8_t msgId   = 'A';
+  uint16_t msgId   = 'A';
   int     msgType = AddOrderLongMsg;
 
   conf.params.bytes4Strategy[msgType].msgId    = msgId;
